added services directory for routes functions and cleaner code
