+++ backend/app.js
@@ -6,9 +6,11 @@ import Auth from './models/Auth.js'
 
 const checkToken = async function (req, res , next) {
 
-    if (req.path === "./signin" || req.path === "./register" ||(req.path === "/projects" && req.method === "GET" || req.path === "/upload")){
+    if (req.path === "/signin" || req.path === "/register" ||(req.path === "/projects" && req.method === "GET" || req.path === "/upload")){
+        console.log("salam");
         next()
     } else {
+        console.log("ssss");
         const token = req.body.token
         const foundedUser = await Auth.findOne({token : token})
         if (foundedUser){
