+++ backend/package.json
@@ -4,6 +4,7 @@
     "body-parser": "^1.20.2",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
+    "lodash": "^4.17.21",
     "mongoose": "^8.1.1",
     "multer": "^1.4.5-lts.1"
   }
